// src/models/DiscountCodeUpdateDTO.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Auth
 * Auth API
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface DiscountCodeUpdateDTO
 */
export interface DiscountCodeUpdateDTO {
    /**
     * 
     * @type {string}
     * @memberof DiscountCodeUpdateDTO
     */
    code: string;
}

export function DiscountCodeUpdateDTOFromJSON(json: any): DiscountCodeUpdateDTO {
    return DiscountCodeUpdateDTOFromJSONTyped(json, false);
}

export function DiscountCodeUpdateDTOFromJSONTyped(json: any, ignoreDiscriminator: boolean): DiscountCodeUpdateDTO {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'code': json['code'],
    };
}

export function DiscountCodeUpdateDTOToJSON(value?: DiscountCodeUpdateDTO | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'code': value.code,
    };
}