// src/models/PlatformMethodUpdateDTO.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Auth
 * Auth API
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import {
    IdReferenceDTO,
    IdReferenceDTOFromJSON,
    IdReferenceDTOFromJSONTyped,
    IdReferenceDTOToJSON,
} from './';

/**
 * 
 * @export
 * @interface PlatformMethodUpdateDTO
 */
export interface PlatformMethodUpdateDTO {
    /**
     * 
     * @type {IdReferenceDTO}
     * @memberof PlatformMethodUpdateDTO
     */
    method: IdReferenceDTO;
}

export function PlatformMethodUpdateDTOFromJSON(json: any): PlatformMethodUpdateDTO {
    return PlatformMethodUpdateDTOFromJSONTyped(json, false);
}

export function PlatformMethodUpdateDTOFromJSONTyped(json: any, ignoreDiscriminator: boolean): PlatformMethodUpdateDTO {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'method': IdReferenceDTOFromJSON(json['method']),
    };
}

export function PlatformMethodUpdateDTOToJSON(value?: PlatformMethodUpdateDTO | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'method': IdReferenceDTOToJSON(value.method),
    };
}